+++ seedfarmer/commands/_module_commands.py
@@ -260,7 +260,7 @@ def _execute_module_commands(
             f"{config.PROJECT.lower()}-{deployment_name}-{group_name}"
             f"-{module_manifest_name}-{generate_session_hash(session=session)}"
         ),
-        bundle_id=module_manifest_name,
+        bundle_id=f"{deployment_name}-{group_name}-{module_manifest_name}",
         codebuild_compute_type=codebuild_compute_type,
         extra_files={config.CONFIG_FILE: os.path.join(config.OPS_ROOT, config.CONFIG_FILE)},
         boto3_session=session_getter,