+++ CHANGELOG.md
@@ -10,6 +10,7 @@ This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a Ch
 - Added intra-group validation of parameter references (prevent any intra-group dependencies)
 
 ### Changes
+- catch exceptions when deleteing a deployment but the project policy (stack) is still in use elewhere
 
 ### Fixes
 - updated pip library to `certifi~=2022.12.7` in requirements-dev (ref dependabot #4)