+++ CHANGELOG.md
@@ -8,6 +8,8 @@ This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a Ch
 
 ### New
 - Added intra-group validation of parameter references (prevent any intra-group dependencies)
+- Added validation of module deletion to prevent deleting modules that have dependencies
+- Added CLI lookup `seedfarmer list dependencies` to provide dependencies of a module
 
 ### Changes
 - catch exceptions when deleting a deployment but the project policy (stack) is still in use elewhere